downgrade hunks as they get passed to SMTP
There is a *horrible* bug in Net::SMTP.  It will encode any input
string if the SvUTF8 flag it turned on.  This is not a sensible
thing to do, and will result in double-encoding binary input based
on the totally irrelevant question of how the runtime is storing
the string elements.

See: https://rt.cpan.org/Ticket/Display.html?id=104433

This fix is stupid, but should work.  We downgrade the string into
UTF-8 octets stored in a non SvUTF8 scalar, which tricks Net::SMTP
into behaving nicely.

I look forward to removing this hack and requiring a newer Net::SMTP
in the future.